Rename space_new.py to space.py
