+++ tests/test_dimensions.py
@@ -13,8 +13,8 @@ os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
 
 import zfit
 from zfit.core.dimension import limits_overlap
-from zfit.core.space_new import combine_spaces
-from zfit.core.space_new import limits_consistent
+from zfit.core.space import combine_spaces
+from zfit.core.space import limits_consistent
 
 obs = ['obs' + str(i) for i in range(4)]
 space1 = zfit.Space(obs=obs)
@@ -84,7 +84,7 @@ def test_combine_spaces():
 
 def test_add_spaces():
     # TODO: add more tests to test the new spaces
-    from zfit.core.space_new import add_spaces_new as add_spaces
+    from zfit.core.space import add_spaces_new as add_spaces
 
     # with pytest.raises(ValueError):
     #     assert add_spaces(space1)
