+++ zfit/core/__init__.py
@@ -1,4 +1,4 @@
 #  Copyright (c) 2020 zfit
 
 from . import (baseobject, basepdf, basemodel, basefunc, data, interfaces, integration, loss,
-               sample, limits, operations, parameter, )
+               sample, space, operations, parameter, )
