+++ zfit/core/basemodel.py
@@ -18,7 +18,7 @@ from .baseobject import BaseNumeric
 from .data import Data, Sampler, SampleData
 from .dimension import BaseDimensional
 from .interfaces import ZfitModel, ZfitParameter, ZfitData, ZfitSpace
-from .limits import Space, convert_to_space, no_norm_range, supports
+from .space import Space, convert_to_space, no_norm_range, supports
 from .sample import UniformSampleAndWeights
 from .. import z
 from ..core.integration import Integration
