--- zfit/core/limits.py
@@ -1,3 +1,0 @@
-#  Copyright (c) 2020 zfit
-
-from .space_new import *
