+++ zfit/models/basefunctor.py
@@ -5,9 +5,9 @@ from typing import List, Union, Tuple
 
 from ..core.basemodel import BaseModel
 from ..core.dimension import get_same_obs
-from ..core.space_new import combine_spaces
+from ..core.space import combine_spaces
 from ..core.interfaces import ZfitFunctorMixin, ZfitModel
-from ..core.limits import Space
+from ..core.space import Space
 from ..util.container import convert_to_container
 from ..util.exception import NormRangeNotSpecifiedError, LimitsIncompatibleError
