+++ zfit/models/dist_tfp.py
@@ -24,9 +24,9 @@ from ..util import ztyping
 from ..settings import ztypes
 from ..core.basepdf import BasePDF
 from ..core.interfaces import ZfitParameter, ZfitData
-from ..core.limits import no_norm_range, supports
+from ..core.space import no_norm_range, supports
 from ..core.parameter import convert_to_parameter
-from ..core.limits import Space
+from ..core.space import Space
 
 
 def tfd_analytic_sample(n: int, dist: tfd.Distribution, limits: ztyping.ObsTypeInput):
@@ -97,7 +97,7 @@ class WrapDistribution(BasePDF):  # TODO: extend functionality of wrapper, like 
     # TODO: register integral
     @supports()
     def _analytic_integrate(self, limits, norm_range):
-        lower, upper = limits.rect_limits_tf
+        lower, upper = limits._rect_limits_tf
         tf.debugging.assert_all_finite((lower, upper), "Are infinite limits needed? Causes troubles with NaNs")
         # if np.all(-np.array(lower) == np.array(upper)) and np.all(np.array(upper) == np.infty):
         #     return z.to_real(1.)  # tfp distributions are normalized to 1
